Перелистывания содержимого БД (#22)
* Реализован функционал перелистывания содержимого БД
* Кнопки страниц автоматически отображаются\убираются в соотвествии с количеством строк в БД
* Убран повтор кода
* Мелкие исправления

+++ CRM/database-types/advertisement.h
@@ -27,8 +27,7 @@ struct advertisement {
 
     advertisement() = default;
 
-    advertisement(int id, std::string source, long long budget,
-                  std::string comments)
+    advertisement(int id, std::string source, int budget, std::string comments)
         : id(id), source(source), budget(budget), comments(comments) {}
 
     advertisement(const advertisement& other)
